Moving the new extension methods to OrchardFeaturesTestingUITestContextExtensions, adding test for workflows
